Feat: adds scripts and updates README
This new homebrew script allows you to execute four commands with just one. For example, instead of doing update and upgrade, you can execute them all together. However, to keep track of what's going on, the respective commands are still output using figlet & Co. Give it a try!

It updates the README as well and creates/initializes the setup script.